fix(rocm): stop three places still describing the pre-#366 page-size contract
Copilot round 7. All three are consequences of correcting the native set to
{1, 16, 1024}, which this PR did and their text predates.

**The standard paged-prefill benchmark rows measured the gather.** The testlist
chose page 128 precisely to avoid the flat gather's in-timed-region copy, and
128 is not native -- it never was; that was the belief this PR disproved. Those
six rows move to 1024, the one size `auto` routes natively for bf16. Verified
they now reach the kernel the header claims:

    fa2   bs1  page1024 ->        0.2628 ms
    auto  bs1  page1024 -> aiter  0.0547 ms
    fa2   bs16 page1024 ->        2.5992 ms
    auto  bs16 page1024 -> aiter  0.5904 ms

**The probe-failure warning promised a fallback fp8 never gets.** The text is
shared, so an fp8 config whose native probe failed was told it was "falling
back to the flat-gather path" immediately before plan() raised because no such
kernel exists. The outcome clause is now dtype-aware, and the docstring's
"falling back is always correct" says which dtypes it is correct for.

**A route check used the wrong dtype.** The gather-fallback numerics test
queried `_aiter_paged_route_page_sizes(torch.float16)` and then planned in
bfloat16; the helper is dtype-dependent by design, so the two diverging would
have made it run or skip on the wrong answer.

Declined: adding the fp8/qscale axes to `flashinfer/jit/rocm/aiter_variants.py`
so the prebuild store covers them. fp8 is batch-prefill only, and that family
is already `servable_from_store=False` -- `_aiter_bootstrap_batch_prefill` is
also the page-size capability probe, so it runs whether or not the .so is in
the store. Prebuilding it fills the store with files nothing ever saves time
on, which is what that flag exists to say.

+++ tests/rocm/test_batch_prefill_kernels.py
@@ -1139,7 +1139,9 @@ def test_batch_prefill_aiter_falls_back_when_native_paging_missing(
     device = torch.device("cuda:0")
     if not is_aiter_supported(device) or not _aiter_ops_importable():
         pytest.skip("AITER requires a gfx942/gfx950 GPU and the aiter package")
-    if page_size not in _aiter_paged_route_page_sizes(torch.float16):
+    # bfloat16: the dtype this test actually plans with, and the route set is
+    # dtype-dependent.
+    if page_size not in _aiter_paged_route_page_sizes(torch.bfloat16):
         pytest.skip(f"page_size={page_size} is not routed natively on this build")
     # Ends in an assert_close against fa2 with causal=True, so it is a numerics
     # test despite being named for the fallback.
